requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
Django==4.2.9
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==2.1
idna==3.6
kombu==5.3.6
prompt-toolkit==3.0.43
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
sparkpost==1.3.10
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
